+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryClient.java
@@ -348,16 +348,28 @@ public class BigQueryClient {
   public Job createAndWaitFor(JobConfiguration jobConfiguration) {
     JobInfo jobInfo = JobInfo.of(jobConfiguration);
     Job job = bigQuery.create(jobInfo);
+    Job returnedJob = null;
 
     log.info("Submitted job {}. jobId: {}", jobConfiguration, job.getJobId());
-    // TODO(davidrab): add retry options
     try {
-      return job.waitFor();
+      Job completedJob = job.waitFor();
+      if (completedJob == null) {
+        throw new BigQueryException(
+            BaseHttpServiceException.UNKNOWN_CODE,
+            String.format("Failed to run the job [%s], got null back", job));
+      }
+      if (completedJob.getStatus().getError() != null) {
+        throw new BigQueryException(
+            BaseHttpServiceException.UNKNOWN_CODE,
+            String.format(
+                "Failed to run the job [%s], due to '%s'", completedJob.getStatus().getError()));
+      }
+      return completedJob;
     } catch (InterruptedException e) {
       Thread.currentThread().interrupt();
       throw new BigQueryException(
           BaseHttpServiceException.UNKNOWN_CODE,
-          String.format("Failed to run the job [%s]", job),
+          String.format("Failed to run the job [%s], task was interrupted", job),
           e);
     }
   }