+++ bigquery-connector-common/src/test/java/com/google/cloud/bigquery/connector/common/BigQueryUtilTest.java
@@ -24,6 +24,7 @@ import com.google.cloud.bigquery.Clustering;
 import com.google.cloud.bigquery.ExternalTableDefinition;
 import com.google.cloud.bigquery.Field;
 import com.google.cloud.bigquery.Field.Mode;
+import com.google.cloud.bigquery.FieldList;
 import com.google.cloud.bigquery.FormatOptions;
 import com.google.cloud.bigquery.HivePartitioningOptions;
 import com.google.cloud.bigquery.LegacySQLTypeName;
@@ -251,30 +252,6 @@ public class BigQueryUtilTest {
   }
 
   @Test
-  public void testFieldWritableScale() {
-    Field f1 = Field.newBuilder("foo", StandardSQLTypeName.INT64).setScale(1L).build();
-    Field f2 = Field.newBuilder("foo", StandardSQLTypeName.INT64).setScale(2L).build();
-    Field f3 = Field.newBuilder("foo", StandardSQLTypeName.INT64).setScale(3L).build();
-    Field f4 = Field.newBuilder("foo", StandardSQLTypeName.INT64).build();
-    assertThat(BigQueryUtil.fieldWritable(f1, f2, true)).isTrue();
-    assertThat(BigQueryUtil.fieldWritable(f3, f2, true)).isFalse();
-    assertThat(BigQueryUtil.fieldWritable(f3, f4, true)).isFalse();
-    assertThat(BigQueryUtil.fieldWritable(f4, f2, true)).isFalse();
-  }
-
-  @Test
-  public void testFieldWritablePrecision() {
-    Field f1 = Field.newBuilder("foo", StandardSQLTypeName.INT64).setPrecision(1L).build();
-    Field f2 = Field.newBuilder("foo", StandardSQLTypeName.INT64).setPrecision(2L).build();
-    Field f3 = Field.newBuilder("foo", StandardSQLTypeName.INT64).setPrecision(3L).build();
-    Field f4 = Field.newBuilder("foo", StandardSQLTypeName.INT64).build();
-    assertThat(BigQueryUtil.fieldWritable(f1, f2, true)).isTrue();
-    assertThat(BigQueryUtil.fieldWritable(f3, f2, true)).isFalse();
-    assertThat(BigQueryUtil.fieldWritable(f3, f4, true)).isFalse();
-    assertThat(BigQueryUtil.fieldWritable(f4, f2, true)).isFalse();
-  }
-
-  @Test
   public void testSchemaWritableWithEnableModeCheckForSchemaFields() {
     Schema s1 =
         Schema.of(
@@ -480,5 +457,124 @@ public class BigQueryUtilTest {
     String tooLarge =
         IntStream.range(0, 2 + 2 << 20).mapToObj(i -> "a").collect(Collectors.joining());
     assertThat(BigQueryUtil.filterLengthInLimit(Optional.of(tooLarge))).isFalse();
+  }
+
+  @Test
+  public void testGetPrecision() throws Exception {
+    assertThat(
+            BigQueryUtil.getPrecision(
+                Field.newBuilder("foo", StandardSQLTypeName.NUMERIC).setPrecision(5L).build()))
+        .isEqualTo(5);
+    assertThat(
+            BigQueryUtil.getPrecision(Field.newBuilder("foo", StandardSQLTypeName.NUMERIC).build()))
+        .isEqualTo(38);
+    assertThat(
+            BigQueryUtil.getPrecision(
+                Field.newBuilder("foo", StandardSQLTypeName.BIGNUMERIC).build()))
+        .isEqualTo(76);
+    assertThat(BigQueryUtil.getPrecision(Field.newBuilder("foo", StandardSQLTypeName.BOOL).build()))
+        .isEqualTo(-1);
+  }
+
+  @Test
+  public void testGetScale() throws Exception {
+    assertThat(
+            BigQueryUtil.getScale(
+                Field.newBuilder("foo", StandardSQLTypeName.NUMERIC).setScale(5L).build()))
+        .isEqualTo(5);
+    assertThat(BigQueryUtil.getScale(Field.newBuilder("foo", StandardSQLTypeName.NUMERIC).build()))
+        .isEqualTo(9);
+    assertThat(
+            BigQueryUtil.getScale(Field.newBuilder("foo", StandardSQLTypeName.BIGNUMERIC).build()))
+        .isEqualTo(38);
+    assertThat(BigQueryUtil.getScale(Field.newBuilder("foo", StandardSQLTypeName.BOOL).build()))
+        .isEqualTo(-1);
+  }
+
+  @Test
+  public void testAdjustSchemaIfNeeded() {
+    Schema wantedSchema =
+        Schema.of(
+            Field.of("numeric", LegacySQLTypeName.NUMERIC),
+            Field.of(
+                "record",
+                LegacySQLTypeName.RECORD,
+                Field.of("subfield", LegacySQLTypeName.STRING)));
+    Schema existingTableSchema =
+        Schema.of(
+            Field.of("numeric", LegacySQLTypeName.BIGNUMERIC),
+            Field.of(
+                "record",
+                LegacySQLTypeName.RECORD,
+                Field.of("subfield", LegacySQLTypeName.STRING)));
+    Schema adjustedSchema = BigQueryUtil.adjustSchemaIfNeeded(wantedSchema, existingTableSchema);
+    assertThat(adjustedSchema.getFields()).hasSize(2);
+    FieldList adjustedFields = adjustedSchema.getFields();
+    assertThat(adjustedFields.get("numeric").getType()).isEqualTo(LegacySQLTypeName.BIGNUMERIC);
+    assertThat(adjustedFields.get("record").getType()).isEqualTo(LegacySQLTypeName.RECORD);
+    assertThat(adjustedFields.get("record").getSubFields()).hasSize(1);
+    assertThat(adjustedFields.get("record").getSubFields().get(0).getType())
+        .isEqualTo(LegacySQLTypeName.STRING);
+  }
+
+  @Test
+  public void testAdjustSchemaForNewField() {
+    Schema wantedSchema =
+        Schema.of(
+            Field.of("existing_field", LegacySQLTypeName.NUMERIC),
+            Field.of("new_field", LegacySQLTypeName.STRING));
+    Schema existingTableSchema =
+        Schema.of(Field.of("existing_field", LegacySQLTypeName.BIGNUMERIC));
+    Schema adjustedSchema = BigQueryUtil.adjustSchemaIfNeeded(wantedSchema, existingTableSchema);
+    assertThat(adjustedSchema.getFields()).hasSize(2);
+    FieldList adjustedFields = adjustedSchema.getFields();
+    assertThat(adjustedFields.get("existing_field").getType())
+        .isEqualTo(LegacySQLTypeName.BIGNUMERIC);
+    assertThat(adjustedFields.get("new_field").getType()).isEqualTo(LegacySQLTypeName.STRING);
+  }
+
+  @Test
+  public void testAdjustField_no_op() {
+    Field field = Field.of("f", LegacySQLTypeName.BOOLEAN);
+    Field existingField = Field.of("f", LegacySQLTypeName.BIGNUMERIC);
+    Field adjustedField = BigQueryUtil.adjustField(field, existingField);
+    assertThat(adjustedField.getType()).isEqualTo(LegacySQLTypeName.BOOLEAN);
+  }
+
+  @Test
+  public void testAdjustField_numeric_to_big_numeric() {
+    Field field = Field.of("numeric", LegacySQLTypeName.NUMERIC);
+    Field existingField = Field.of("numeric", LegacySQLTypeName.BIGNUMERIC);
+    Field adjustedField = BigQueryUtil.adjustField(field, existingField);
+    assertThat(adjustedField.getType()).isEqualTo(LegacySQLTypeName.BIGNUMERIC);
+  }
+
+  @Test
+  public void testAdjustFieldRecursive() {
+    Field field =
+        Field.of(
+            "record", LegacySQLTypeName.RECORD, Field.of("subfield", LegacySQLTypeName.STRING));
+    Field existingField =
+        Field.of(
+            "record", LegacySQLTypeName.RECORD, Field.of("subfield", LegacySQLTypeName.STRING));
+    Field adjustedField = BigQueryUtil.adjustField(field, existingField);
+    assertThat(adjustedField.getType()).isEqualTo(LegacySQLTypeName.RECORD);
+    assertThat(adjustedField.getSubFields()).hasSize(1);
+    assertThat(adjustedField.getSubFields().get(0).getType()).isEqualTo(LegacySQLTypeName.STRING);
+  }
+
+  @Test
+  public void testAdjustFieldRecursive_with_bignumeric_conversion() {
+    Field field =
+        Field.of(
+            "record", LegacySQLTypeName.RECORD, Field.of("subfield", LegacySQLTypeName.NUMERIC));
+    Field existingField =
+        Field.of(
+            "record", LegacySQLTypeName.RECORD, Field.of("subfield", LegacySQLTypeName.BIGNUMERIC));
+    Field adjustedField = BigQueryUtil.adjustField(field, existingField);
+    assertThat(adjustedField.getType()).isEqualTo(LegacySQLTypeName.RECORD);
+    assertThat(adjustedField.getSubFields()).hasSize(1);
+    assertThat(adjustedField.getSubFields().get(0).getType())
+        .isEqualTo(LegacySQLTypeName.BIGNUMERIC);
   }
 }