+++ spark-bigquery-connector-common/src/main/java/org/apache/spark/bigquery/BigQueryDataTypes.java
@@ -1,7 +1,0 @@
-package org.apache.spark.bigquery;
-
-import org.apache.spark.sql.types.DataType;
-
-public class BigQueryDataTypes {
-  public static final DataType BigNumericType = new BigNumericUDT();
-}