+++ spark-bigquery-connector-common/src/test/java/com/google/cloud/spark/bigquery/integration/SparkBigQueryIntegrationTestBase.java
@@ -66,8 +66,7 @@ public class SparkBigQueryIntegrationTestBase {
               TestConstants.ALL_TYPES_TABLE_QUERY_TEMPLATE,
               testDataset,
               TestConstants.ALL_TYPES_TABLE_NAME));
-      IntegrationTestUtils.createView(
-          testDataset, TestConstants.ALL_TYPES_TABLE_NAME, TestConstants.ALL_TYPES_VIEW_NAME);
+      IntegrationTestUtils.createView(testDataset, TestConstants.ALL_TYPES_VIEW_NAME);
       IntegrationTestUtils.runQuery(
           String.format(
               TestConstants.STRUCT_COLUMN_ORDER_TEST_TABLE_QUERY_TEMPLATE,